Created event triggering/listening support through socket messages
While ChuckT does not currently have a hard dependency on SockJS, it was
designed with the stable SockJS API in mind. Events can be emitted to the
frontend by calling chuckt.emit(). You can listen for events that are fired
from the frontend by calling chuckt.on().

It is assumed that the ChuckT client library will be used on the frontend.